healthcare-ml-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pymongo>=4.6.0
numpy>=2.0.0  # First version with 3.13 support
scikit-learn>=1.5.0  # Requires numpy>=2.0.0
mlflow>=2.10.1
python-dotenv>=1.0.0
pydantic>=2.5.3
pandas>=2.2.0  # Requires numpy>=2.0.0